refactor(daemon): split perf_bench_test.rs into focused modules
Convert tests/perf_bench_test.rs (4,238 lines) into a perf_bench/
module directory with 14 focused submodules. AI tools and reviewers
can now navigate the perf benchmark code without loading the whole
file. No behavioral changes — purely a code split.

Test discovery is preserved: tests/perf_bench_test.rs is now a thin
shim that #[path]-includes perf_bench/mod.rs, so cargo still emits
a single perf_bench_test binary. Substring filters from the canonical
`--test perf_bench_test -- <name>` invocation still match (now via
the module-prefixed `perf_bench::tests_X::<name>` paths), keeping
ci/perf_guard.py, ci/benchmark_stats.py, and perf-guard.yml working
without changes.

Layout:
- mod.rs: module declarations
- common.rs: start_daemon, sccache/em++/archiver finders, timing
  helpers, shared tool runners, session helpers, constants,
  ClientConn type alias
- c_project.rs: C source generation + C compile helpers
- cpp_project.rs: C++ source generation + warmup + single/multi
  compile helpers
- response_file.rs: response-file generation + _rsp compile variants
- rust_project.rs: Rust source generation + rustc batch runners
- link.rs: LinkBenchResult, measure_ephemeral_link_scenario,
  print_link_benchmark_table + link input preparation
- sibling_remap.rs: make_git_workspace, path_remap_auto_env,
  measure_cpp_sibling_remap_mode
- tests_c.rs: perf_c_zccache_vs_bare,
  generated_c_project_compiles_under_std_c11
- tests_cpp.rs: perf_warm_cache_zccache_vs_sccache
- tests_response_file.rs: perf_response_file
- tests_rust.rs: perf_rustc_zccache_vs_sccache
- tests_sibling_remap.rs: perf_cpp_sibling_remap_warm,
  perf_rustc_sibling_remap_warm
- tests_emcc.rs: perf_emcc_warm_cache_zccache_vs_sccache,
  perf_emcc_sibling_remap_warm
- tests_link.rs: perf_c_archive_link, perf_cpp_driver_link,
  perf_emcc_link, perf_rust_workspace_link

After the split, every file is under 500 LOC (largest is link.rs
at 461). The thin shim perf_bench_test.rs is 13 LOC.

Verified:
- soldr cargo check --workspace --all-targets: ok
- soldr cargo clippy --workspace --all-targets -- -D warnings: ok
- soldr cargo test -p zccache-daemon --lib: 226 passed
- All 12 perf_ tests + generated_c_project_compiles_under_std_c11
  are discoverable via the test binary; every existing test-name
  filter resolves to exactly one test.
